+++ src/Components/Notifications/ShowPushNotification.tsx
@@ -1,4 +1,4 @@
-import { DetailRoute } from "../../Routers/types";
+import DetailRoute from "../../Routers/types/DetailRoute";
 import useQuery from "../../Utils/request/useQuery";
 import routes from "../../Redux/api";
 import { NotificationData } from "./models";
