+++ src/Routers/routes/PatientRoutes.tsx
@@ -3,7 +3,7 @@ import { PatientManager } from "../../Components/Patient/ManagePatients";
 import { PatientHome } from "../../Components/Patient/PatientHome";
 import PatientNotes from "../../Components/Patient/PatientNotes";
 import { PatientRegister } from "../../Components/Patient/PatientRegister";
-import { DetailRoute } from "../types";
+import DetailRoute from "../types/DetailRoute";
 import DeathReport from "../../Components/DeathReport/DeathReport";
 import { InsuranceDetails } from "../../Components/Patient/InsuranceDetails";
 import FileUploadPage from "../../Components/Patient/FileUploadPage";
