+++ src/Routers/routes/ResourceRoutes.tsx
@@ -5,7 +5,7 @@ import { ResourceDetailsUpdate } from "../../Components/Resource/ResourceDetails
 import ListView from "../../Components/Resource/ListView";
 import BoardView from "../../Components/Resource/ResourceBoardView";
 import { Redirect } from "raviger";
-import { DetailRoute } from "../types";
+import DetailRoute from "../types/DetailRoute";
 
 const getDefaultView = () =>
   localStorage.getItem("defaultResourceView") === "list" ? "list" : "board";
