+++ src/Routers/routes/SampleRoutes.tsx
@@ -2,7 +2,8 @@ import { SampleDetails } from "../../Components/Patient/SampleDetails";
 import SampleReport from "../../Components/Patient/SamplePreview";
 import { SampleTest } from "../../Components/Patient/SampleTest";
 import SampleViewAdmin from "../../Components/Patient/SampleViewAdmin";
-import { DetailRoute, RouteParams } from "../types";
+import DetailRoute from "../types/DetailRoute";
+import RouteParams from "../types/RouteParams";
 
 export default {
   "/sample": () => <SampleViewAdmin />,
