--- src/Routers/types.ts
@@ -1,5 +1,0 @@
-export type RouteParams<T extends string> = Record<T, string>;
-
-export interface DetailRoute {
-  id: string;
-}
