+++ src/Routers/types/DetailRoute.ts
@@ -1,0 +1,3 @@
+export default interface DetailRoute {
+  id: string;
+}
