+++ src/Routers/types/RouteParams.ts
@@ -1,0 +1,3 @@
+type RouteParams<T extends string> = Record<T, string>;
+
+export default RouteParams;
